Move exhibition prismic stuff to prismic service
