+++ events/app/controllers.js
@@ -1,8 +1,6 @@
-import {Prismic, services, prismic, model, prismicParsers} from 'common';
-
-const {prismicImage, asText} = prismicParsers;
-const {getPrismicApi} = services;
+import {model, prismic} from 'common';
 const {createPageConfig} = model;
+const {getPaginatedResults} = prismic;
 
 export async function renderEvent(ctx, next, overrideId, gaExp) {
   const id = overrideId || `${ctx.params.id}`;
@@ -46,53 +44,20 @@ export async function renderEvent(ctx, next, overrideId, gaExp) {
   return next();
 }
 
-async function getEvents() {
-  const prismic = await getPrismicApi();
-  const eventsList = await prismic.query([
-    Prismic.Predicates.any('document.type', ['events'])
-  ]);
-  return eventsList;
-}
-
-function getNumberFromString(string: string): number {
-  return Number(string.replace(/\D/g, ''));
-}
-
 export async function renderEventsList(ctx, next) {
-  const events = await getEvents();
-  const path = ctx.request.url;
-  const allEvents = events.results.map(event => {
-    const promo = event.data.promo[0];
-    const promoImage = promo && promo.primary.image;
-    const promoCaption = promo && promo.primary.caption;
-
-    return event.data.times.map(eventAtTime => {
-      return {
-        id: event.id,
-        title: asText(event.data.title),
-        url: `/events/${event.id}`,
-        start: eventAtTime.startDateTime,
-        end: eventAtTime.endDateTime,
-        image: prismicImage(promoImage),
-        description: asText(promoCaption)
-      };
-    });
-  }).reduce((acc, curr) => {
-    return curr.concat(acc);
-  }, []).sort((a, b) => {
-    return getNumberFromString(b.start || '') - getNumberFromString(a.start || '');
-  });
+  const page = Number(ctx.request.query.page);
+  const eventsList = await getPaginatedResults(page, 'event');
 
   ctx.render('pages/events', {
     pageConfig: createPageConfig({
-      path: path,
+      path: ctx.request.url,
       title: 'Events',
       inSection: 'whatson',
       category: 'list', // TODO: update to team (ev&ex)
       contentType: 'event', // TODO: add pageType (list)
       canonicalUri: '/events'
     }),
-    allEvents: allEvents
+    eventsList
   });
 
   return next();
