+++ server/views/components/list-header/list-header.njk
@@ -1,0 +1,23 @@
+<div class="row bg-cream  {{ {s:3, m:5, l:5} | spacingClasses({padding: ['top']}) }} }}">
+  <div class="container">
+    <div class="grid">
+      <div class="grid__cell">
+        <h1 class="{{ {s:'WB6', m:'WB5', l:'WB4'} | fontClasses }} {{ {s:0, m:0, l:0} | spacingClasses({margin: ['top']}) }} {{ {s:0, m:0, l:0} | spacingClasses({margin: ['bottom']}) }}">{{ model.name }}</h1>
+      </div>
+    </div>
+
+    {% if model.description %}
+      <div class="grid">
+        <div class="{{ {s: 12, m: 12, l: 8} | gridClasses }}">
+          <p class="{{ {s:2, m:2, l:2} | spacingClasses({margin: ['top']}) }} {{ {s:0, m:0, l:0} | spacingClasses({margin: ['bottom']}) }}">{{ model.description }}</p>
+        </div>
+      </div>
+    {% endif %}
+
+    <div class="grid">
+      <div class="grid__cell">
+        <p class="font-pewter {{ {s:'LR3', m:'LR2'} | fontClasses }} {{ {s:2, m:2, l:2} | spacingClasses({margin: ['top']}) }} {{ {s:2, m:2, l:2} | spacingClasses({margin: ['bottom']}) }}">{{ model.total }} items</p>
+      </div>
+    </div>
+  </div>
+</div>
